chore: remove leftover comment
