+++ src/components/Content/DashboardContent.ts
@@ -1,0 +1,6 @@
+export class DashboardContent {
+  public element: HTMLElement;
+  constructor(container: HTMLElement) {
+    this.element = container;
+  }
+}
