+++ src/components/Content/ResultsContent.ts
@@ -1,0 +1,6 @@
+export class ResultsContent {
+  public element: HTMLElement;
+  constructor(container: HTMLElement) {
+    this.element = container;
+  }
+}
