+++ src/components/Content/SuggestContent.ts
@@ -1,0 +1,6 @@
+export class SuggestContent {
+  public element: HTMLElement;
+  constructor(container: HTMLElement) {
+    this.element = container;
+  }
+}
