+++ src/components/Content/index.ts
@@ -1,0 +1,3 @@
+export * from './DashboardContent';
+export * from './SuggestContent';
+export * from './ResultsContent';
