+++ src/services/index.ts
@@ -12,3 +12,6 @@ export { MapSyncService, MapSyncServiceFactory } from './MapSyncService';
 
 // Экспорт менеджера карты
 export { MapManager } from './MapManager';
+
+// Экспорт менеджера контента
+export { ContentManager } from './ContentManager';
